Fixed previous problems
The design of the Db was flawed for what it had to do

+++ TestAdeverinta/TestDb/dbo/Tables/Certificate.sql
@@ -4,7 +4,7 @@
     [Year] NCHAR(10) NOT NULL, 
     [Volume] INT NOT NULL, 
     [AddressID] INT NOT NULL,
-    [PersonCNP] INT NOT NULL, 
+    [PersonCNP] BIGINT NOT NULL, 
     FOREIGN KEY (PersonCNP) REFERENCES PersonTable(CNP),
     FOREIGN KEY (AddressID) REFERENCES AddressTable(Id),
 );
